Pull request #5095: extractor: avoid reporting default values for missing SSL fields
Merge in SNORT/snort3 from ~YCHALOV/snort3:ssl_extractor_default_value to master

Squashed commit of the following:

commit 938939d2d569882eb913b7fc1870eb722900b9a5
Author: Yurii Chalov <ychalov@cisco.com>
Date:   Wed Jan 14 11:56:48 2026 +0100

    extractor: avoid reporting default values for missing SSL fields

+++ src/service_inspectors/ssl/ssl_inspector.cc
@@ -108,7 +108,7 @@ private:
 
 const std::string Ssl::s_name = "ssl";
 
-uint16_t SslMetadataEvent::get_version() const
+int32_t SslMetadataEvent::get_version() const
 {
     return tls_connection_data.tls_params.selected_tls_version;
 }
@@ -118,12 +118,12 @@ const std::string& SslMetadataEvent::get_server_name_identifier() const
     return tls_connection_data.server_name_identifier;
 }
 
-uint16_t SslMetadataEvent::get_curve() const
+int32_t SslMetadataEvent::get_curve() const
 {
     return tls_connection_data.tls_params.curve;
 }
 
-uint16_t SslMetadataEvent::get_cipher() const
+int32_t SslMetadataEvent::get_cipher() const
 {
     return tls_connection_data.tls_params.cipher;
 }
